refactor: updated pr test file and pitr example
