+++ module-metadata.json
@@ -153,7 +153,7 @@
     "pitr_id": {
       "name": "pitr_id",
       "type": "string",
-      "description": "(Optional) The ID of the source deployment that you want to recover back to.",
+      "description": "(Optional) The ID of the postgresql instance that you want to recover back to. Here ID of the postgresql instance is expected to be up and in running state.",
       "source": [
         "ibm_database.postgresql_db.point_in_time_recovery_deployment_id"
       ],
