+++ tests/pr_test.go
@@ -2,10 +2,12 @@
 package test
 
 import (
-	"testing"
-
 	"github.com/stretchr/testify/assert"
+	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/common"
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/testhelper"
+	"log"
+	"os"
+	"testing"
 )
 
 // Use existing resource group
@@ -14,9 +16,12 @@ const defaultExampleTerraformDir = "examples/default"
 const autoscaleExampleTerraformDir = "examples/autoscale"
 const fsCloudTerraformDir = "examples/fscloud"
 const completeExampleTerraformDir = "examples/complete"
+const replicaExampleTerraformDir = "examples/replica"
 
 // Restricting due to limited availability of BYOK in certain regions
 const regionSelectionPath = "../common-dev-assets/common-go-assets/icd-region-prefs.yaml"
+
+const yamlLocation = "../common-dev-assets/common-go-assets/common-permanent-resources.yaml"
 
 func TestRunDefaultExample(t *testing.T) {
 	t.Parallel()
@@ -40,6 +45,21 @@ func TestRunAutoscaleExample(t *testing.T) {
 		Testing:       t,
 		TerraformDir:  autoscaleExampleTerraformDir,
 		Prefix:        "pg-autoscale",
+		ResourceGroup: resourceGroup,
+	})
+
+	output, err := options.RunTestConsistency()
+	assert.Nil(t, err, "This should not have errored")
+	assert.NotNil(t, output, "Expected some output")
+}
+
+func TestRunReplicaExample(t *testing.T) {
+	t.Parallel()
+
+	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
+		Testing:       t,
+		TerraformDir:  replicaExampleTerraformDir,
+		Prefix:        "pg-replica",
 		ResourceGroup: resourceGroup,
 	})
 
@@ -102,4 +122,36 @@ func TestRunUpgradeExample(t *testing.T) {
 		assert.Nil(t, err, "This should not have errored")
 		assert.NotNil(t, output, "Expected some output")
 	}
+}
+
+var permanentResources map[string]interface{}
+
+// TestMain will be run before any parallel tests, used to read data from yaml for use with tests
+func TestMain(m *testing.M) {
+
+	var err error
+	permanentResources, err = common.LoadMapFromYaml(yamlLocation)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	os.Exit(m.Run())
+}
+
+func TestRunPointInTimeRecoveryDBExample(t *testing.T) {
+	t.Parallel()
+
+	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
+		Testing:       t,
+		TerraformDir:  PITRTerraformDir,
+		Prefix:        "pg-pitr",
+		ResourceGroup: resourceGroup,
+		TerraformVars: map[string]interface{}{
+			"pitr_id": permanentResources["postgresqlCrn"],
+		},
+	})
+
+	output, err := options.RunTestConsistency()
+	assert.Nil(t, err, "This should not have errored")
+	assert.NotNil(t, output, "Expected some output")
 }
